practice_age-classification/requirements.txt
transformers
pillow
streamlit

# https://pytorch.org/
torch
torchvision
torchaudio

numpy<2